change for better API for UCBV
